Fix the issue that load push statues will failed while we are restarting controllers.
The scenaior is if we stopped controller for any reason e.g. upgrading, in a short period we don't have master controller for the cluster. During that short period, storage nodes update its replica status which will change the push status e.g. get enough error or online replicas so push should be changed to COMPLETED or ERROR, as there is not master controller, so push status has not been changed. After a new master is elected, it will load all pushes and try to update push statues if needed during the state transtiion STANDBY->LEADER. But it will failed because it has not been the master controller so it could not delete error/completed version.
This RB will catch the exception and log it instead of failing the load all pushes operation, thus new controller will become master successfully.

RB=1104018
G=voldemort-reviewers
R=sdwu
A=fvillega