Fix the issue that controller start up could fail due to incorrect offline push status in ZK.
When version cration fail, we delete that version but do NOT mark the
offline push status as ERROR. Then when we re-start the controller, it
will try to laod all offline push status from ZK, and if the status is
still STARTED, controller will try to fetch the latest replica status
and update the offline push status. But as the version has not been
created yet, an exception will be thrown to say there is not
partition/replica for this version. This will fail the helix transition
from standby to leader. So it means no controller could become the
master controller eventually.
This code change do two things:
1. Mark the offline push as ERROR once version creation fail.
2. Check whether version exists not not before fechting the
replica/partition info for this version during loading offline push
status from zk.

RB=1124582
G=voldemort-reviewers
R=goliu
A=fvillega