Added config to disable topic deletion in Parent Controller
Recently, topic deletion in Parent Controller has caused a lot of
troubles in Kafka MM, since the old consumer has the global re-balance
issue and the new consumer is not be able to handle topic deletion
properly.
The new strategy is to truncate the Kafka topic in Parent controller
when job finishes.
To check whether there are parallel pushes or not, Parent Controller
relies on whether the retention policy of the latest topic is 0 or not.
If not, Parent Controller will query against prod controllers to get the
actual job status.

TODO:
Right now, there is no scheduled topic deletion logic in Parent Controller
yet, and we need to work with Kafka team about the deletion strategy,
which could apply to Prod Controller as well since Prod Controller is
not deleting topics when job fails, which leaves a lot of garbage topics
in Prod.

RB=1133998
G=voldemort-reviewers
R=fvillega
A=fvillega