Improve flaky venice-thin-client tests
These changes add a couple of retries to schema queries, and skip creation of certificates
if they aren't needed.  This resolved some of the flaky tests I encountered but not all of
them. There's still some issue I haven't been able to trace regarding the mock http server
not responding to requests once in a while.

RB=1135368
G=voldemort-reviewers
A=fvillega