Introduced a new StatsErrorCode enum.
This enum tracks the error codes that we use to report anomalies in metrics.

Background: there are various edge cases, both intentional and unintentional,
that can cause us to not emit a metric, either because we don't want to or
because we cannot. Historically, we had many such edge cases that would make
a metric default to zero. This is problematic when debugging why a metric is
not acting right, since we cannot easily disambiguate between the various edge
cases.

In order to ease the debugging of metric anomalies, we are using this enum to
track which code paths emit which error code. This ensures that we don't use
the same sentinel values for two different meanings.

Unrelated main code changes:
1. Tweaked the RatioStat API so that it takes time into account.
2. Tweaked the KafkaStoreIngestionService#startConsumption function so that
   it is resilient to stale metadataRepo information. Also added a log when
   the stale metadata is detected.

Unrelated test code change:
1. Marked a new test as flaky: TestAdminSparkServerWithMultiServers#
   requestTopicIsIdempotentWithConcurrency

RB=1143355
BUG=VOLDENG-3543
G=voldemort-reviewers
R=sdwu
A=sdwu