Catch TopicAlreadyMarkedForDeletion exception and swallow it so jobs can retry the topic delete blocking method
Previously we assumed that kafka would let us try to delete a topic if it exists.  However if the topic is already
marked for deletion then the topic will still exist but it will throw an exception if we try to delete it. This
makes the topic deletion method more idempotent in case of failure

RB=1150263
G=voldemort-reviewers
A=fvillega