Added a couple of new schemas to enable value-chunking.
- ChunkedValueManifest is intended to be stored in the value, and distinguished
  from regular values by tagging it as schema ID -1. It contains various pieces
  of information needed to reconstruct the multi-chunk value.
- ChunkId is intended to prepend the key for record chunks. Whether to use
  this ChunkId as a key prefix or not is going to be decided at the granularity
  of store-version. i.e.: for a store-version that enables this feature, all
  of its keys are going to be prepended. The record itself is optional, which
  means the overhead should be minimal for non-chunked keys.

The two new schemas are placed inside venice-common, rather than venice-server
because we intend to re-use the same GUID class which is part of the Kafka
protocol schemas, and it is easier to achieve this within a single module. They
will also be leveraged from within the VenicePushJob.

In addition to the above new schemas, the following old schemas have also been
updated:

- KafkaMessageEnvelope, v4: Includes a new "chunked" boolean field in the
  StartOfPush control message.
- StoreVersionState, v2: Includes a new "chunked" boolean field as well.

Both of these new fields default to false, if absent.

RB=1163273
G=voldemort-reviewers
R=goliu
A=goliu