off-heap cache in Router
This change is using ohc as the off-heap cache library. Accoring to
test comparison with on-heap cache, the GC impact is minimal,
and the overall performance looks good.
Here is the test result:
https://microsoft-my.sharepoint.com/:w:/g/personal/goliu_linkedin_biz/EWQvNbNrd5hKngYpq3ij06kBoychFEMDehugInCsyXBmog?e=sOKYuC

TODO:
Need to explore how to preload jemalloc to optimize memory
fragementation.

RB=1212659
G=voldemort-reviewers
R=mwise,fvillega,chgao,yayan
A=yayan