Added the ability to alter the largest used version number.
This is useful in certain data loss scenario where we want to perform
a recovery operation by re-copying a topic from a healthy datacenter.

The new command only works on child controllers, not on the parent
controller, and therefore does not touch the admin topic protocol.

Also included is a refactoring of several pieces of admin tool and
controller code, to eliminate redundant boilerplate. Adding new params
to the update store command should require a little bit less code in
the future.

RB=1204730
G=voldemort-reviewers
R=yayan
A=yayan