Exposed schema API through store client
For now, the exposed 'getLatestValueSchema' could not guarantee it will
always return the latest value schema stored in Venice.
It will become a little complicate if we want to return the absolute
latest value schema.
One way is to query Venice Router every time, and it is too expensive
for some user case, such as Samza SQL integration;
Another way is to maintain a async thread to periodically refresh local
schema cache; But it still couldn't guarantee the absolute latest value
schema every time, but eventually it could return it.

For now, Samza team is fine with the current implementation, and customer
should restart their Samza Job when a new value schema is added.

RB=1223758
G=voldemort-reviewers
R=fvillega
A=fvillega