Add some resiliance to the query job status endpoint
We ran into issues where the getLatestOffsets() method on Kafka would
timeout.  This adds the ability for the controller to
continue reporting job status even if that query fails.  It means we
wont have offsets to report with job status, but the job can continue
without failing H2V.  It also adds an ability to query for offsets with
retries, and has the storage node use that when calculating hybrid
status

RB=1234047
G=voldemort-reviewers
A=fvillega,yayan