Refactored the concurrency code in the StorageExecutionHandler.
The new code has less concurrency logic, and it is structured
differently. In particular:

- The main function, #channelRead, is now responsible for kicking
  off the executor submission as soon as it is invoked. Everything
  else in the class is then rewritten as blocking code.
- For single gets, this should eliminate one context switch.
- For batch gets, this removes a large chunk of the code from the
  blocking Netty hot path.

In micro-benchmarking, this seems to improve the latency of highly
concurrent batch get workloads with large numbers of keys by about
30%, though it remains to be seen if this translates into tangible
benefits on the real environment.

There is a new test, for the micro-benchmark, in:

TestBatch#stress#stressTestLargeMultiGet

But it is disabled, since it does not perform any functional testing,
and the automated test suite is likely not the right vehicule for
performance regression testing.

Added a new metric called:

- storage_execution_handler_submission_wait_time: Indicates the wait
  time spent in the StorageExecutionHandler's queue. A high number
  would indicate that there is not enough thread in the executor to
  handle all the work.

Unrelated test changes:

- Stabilized TestDeleteStoreDeletesRealtimeTopic to take into account
  the new async TopicCleanupService.
- AggVersionedDIVStatsTest#testStatsCanLoadAllStoresInTime was marked
  as flaky.
- TestCachedDnsResolver#testRefreshCacheEntries was marked as flaky.
- Pinned the version of Kafka in venice-test-common as Brooklin was
  upgrading it transitively, and that caused some issues.

RB=1279825
BUG=VOLDENG-3873
G=voldemort-reviewers
R=goliu
A=mwise,goliu