Long tail retry for batch-get
This change is used to enable long-tail retry for batch-get.
Based on current design, the long-tail retry threshold is actually based
on the key number in a batch-get request, so here introduces the
following config:
router.long.tail.retry.for.batch.get.threshold.ms

After this amount of time, DDS Router will retry once for the slow
storage node request.

The configured format will be like this way:
"1-10:20,11-50:50,51-200:80,201-:1000"

Let me explain the config by taking one example:
If the request key count for a batch-get request is '32', and it
fails into this key range: [11-50], so the retry
threshold for this batch-get request is 50ms.

That is a limitation here:
The retry threshold is actually for each scatter-gather request,
but this config is not strictly with the actual key count
inside each scatter-gather request, which means even if there is
only one key in a scatter-gather request with
the above example, Router will wait for 50ms to retry this
scatter-gather request.

For now, it is not big issue since for now we mostly want to use
this config to skip the storage node, which
is experiencing long GC pause.
So coarse-grained config should be good enough.

This change also supports retry metrics per request type
(single-get/batch-get);

Another change is to skip quota enforcement for retry request, and the
consideration behind this is that long-tail retry is one way of latency guarantee
offered by Venice, so custemers shouldn't be aware of it.

RB=1243770
G=voldemort-reviewers
R=mwise,sdwu,fvillega,chgao,yayan
A=yayan