Pulled message digest object into thread local for reusability
In performance testing I found that if the storage node is fast (by
using an in-memory storage engine) then under heavy load the router
netty threads spend a lot of time blocked trying to get a message
digest object from the java security packages.  This object can be
reused, but has some state so I persist that object as a ThreadLocal
so it can be reused.  With 100 concurrent batchget requests of 2000
keys each and an in-memory storage engine this improves peformance
(in local unit testing) from an average of 73ms per request to an
average of 46ms per request.

RB=1289907
G=voldemort-reviewers
A=goliu