Version swap race condition fix
There was a race condition in the router for version swap where the router
would see the version swap message before seeing replicas online for every
partition of the new version. This fix adds an extra check in the router
to prevent version swap before it sees online replicas for every partition.

RB=1304879
G=voldemort-reviewers
A=fvillega