Updated multiget request to storage node to disable chunked transfer encoding
During the local test, I realized that multiget request to storage node
is using 'chunked transfer-encoding', which is not expected.
This 'chunked transfer-encoding' will cause the storage node to handle
the request chunk by chunk, which is possibly to introduce some
inefficiency, and this might be the root cause of the latency spike of
multiget request in storage node, since the debugging metrics were
showing the extended latency of handling multiple chunks during multiget
latency spike in storage node.

RB=1302701
BUG=VOLDENG-3929
G=voldemort-reviewers
R=mwise,fvillega
A=fvillega