Catch _v0 does not exist errors and provide a more useful message to partners
In this change we no longer treat 'resource does not exist' as a generic VeniceException.  Instead
we have a new Exception: VeniceNoHelixResourceException, which we catch in the router to provide
a more useful error message and an http 400 if the missing version is version 0.
Note that when we truncate (delete all version from) a store, we set the current version to 0.

RB=1308460
G=voldemort-reviewers
A=fvillega