Fixed race condition during topic deletion in child controller
Previously, the topic deletion steps when retiring a store version:
1. Drop Helix resource (asynchronous);
2. Truncate topic;
3. Delete topic (asynchronous);

So it is possible that the topic gets deleted before the resource is
fully removed when state transition is slow, which means the consumer in
storage node could be still consuming the topic after the topic gets
deleted by {@link TopicCleanupService}, which triggers the non-existing
topic issue of Kafka consumer, which will cause more issues after.

The fix is to check whether the resource is fully removed or not by
checking ExternalView (a little hacky) before deleting the topic to
avoid the race condition mentioned above.

RB=1315075
BUG=VOLDENG-3972
G=voldemort-reviewers
R=fvillega,yayan
A=fvillega