[controller] Fixed a race condition that could stall quick push jobs.
The OfflinePushMonitor could go back to STARTED after COMPLETED, when
quick-running jobs completed very fast. This would cause the push job
to stall forever.

The fix is fairly simple: we need to refresh the offline push before
updating it, to ensure that the transition we are applying is valid.

This regression was introduced in:

7e8030dda3bd54742fa57973373499ffdccbcb2d

RB=1326826
BUG=VOLDENG-3871
G=voldemort-reviewers
R=sdwu,chgao,yayan
A=yayan