Update the store gravyard to support getting the largest used version number from all clusters.
We encountered the push failure because the store was created in cluster1 and deleted. Now it's being created in cluster2, as store gravyard is cluster specific so in the cluster2 controller still use version 1 as the first version but actually it's larger than 1.
Update: Instead of getting the largest used version number from the current cluster, controller will check all cluster's gravyard to get the largest one.

RB=1356859
G=voldemort-reviewers
A=goliu