Fix the issue that the conterller mastership hand-off would fail the ongoing push jobs.
Background: When the master controller creates a new vesion, it has to create couple of item like kafka topic, version, offline push status, helix resource.
Unfortunately it's not an atomic operation, so the thing is the controller might not complete the full process before handing off the mastership.
Once the new master controller takes over the cluster, there is no way to track and complete the process as well. In that case, the push job would faill in the prod colo.
Parent controller does not have this issue, because if a parent controller lose the mastership, the H2V will get the error response and retry.

This change is trying to fix this issue by adding a read-write lock to prevent the mastership hand over and the metadata operation are executed in parallel. So we either get rid of the mastership before create the version or after to avoid partial metadata in zk.
1. Add a read-write lock in venice helix resource
2. Before adding a new version, lock the metadata operation lock(read lock). And unlock once complete the operation.
3. Before stopping the controller, lock the shutdown lock(write lock). And unlock once complete the operation.
4. Before handing off the mastership. lock the shutdown lock(write lock). And unlock once complete the operation.

RB=1360363
G=voldemort-reviewers
R=goliu
A=goliu