[client] Add new exception for http 429, allow client to process any 2xx response
Added a new exception class specifically for throttling.  This allows client applications
to catch the new exception type and handle rate limiting in a different way.  There is
also a change to allow processing of any 2xx response, previously anything other than
a 200 would be treated as an error.

RB=1367538
G=voldemort-reviewers
A=chgao