[server] do BDB checkpoint regularly instead of right after a partition drop
The previous fix would do a checkpoint right after a partition drop, which could bring potential overhead to the system;
now, a checkpoint is done regularly and the behaviour is configurable. A background thread which is responsible for this check
is spawned for each venice server.

bdb.dropped.db.clean.up.enabled=true/false which is used to enable/disable this feature;
server.leaked.resource.clean.up.interval.in.hours=6 which means do checkpoint if necessary every 6 hours

RB=1364673
G=voldemort-reviewers
R=goliu
A=goliu