[router] Config to switch non sticky multiget from greedy to distributed
Without sticky-routing, Venice uses a greedy algorithm to assign partitions to hosts, this results
in an uneven load on the storage nodes--most keys get allocated to a very small number of hosts.
With this change, we add a config for the ScatterGatherMode for multiget to just group by partition.
Instead of using a greedy algorithm it can randomly select a replica for each partition, then group the
replicas on a single host into a single request.

RB=1368108
G=voldemort-reviewers
A=goliu