Fix the SN bounce failure due to storage engine stats
After we moved BdbStorageEngineStats as Versioned stats, it started to
listen to ZK store metadata repository when creating/removing new stats.
If a store is removed during SN is in offline, there is an edge case
when SN rebounces and scans the disk to restore exisiting storage
engine. Since the store is removed in ZK, SN will throw exception and
fail to rebounce.

Catch the exception in the SN to prevent it.

Jira: VOLDENG-4056

RB=1371882
G=voldemort-reviewers
R=mwise,yayan,chgao,goliu
A=mwise