[router] Don't serve old version unless it is online, and check version against correct partition count
The version swap race condition fix (rb 1304879) has an edge case where if a new version has offline
partitions for the duration of multiple new pushes then the router will continue to serve an old
version which is no longer present. This change adds a check to make sure the old version is still online
and if the old version is NOT online then switch over to serving the new version anyways.  There is also
a bug where a version may have a different number of partitions than the store.  This would consistently
under report online partitions.

RB=1379373
G=voldemort-reviewers
A=goliu