[router] Add a new metric to track stale versions
The new metric will show the greatest difference between current version and
serving version across all stores. If it is elevated and stays elevated we'll
need to look at the router logs to identify which store is serving a stale
version.  The metric should be used for alerting and the logs used for debuging

RB=1388560
G=voldemort-reviewers
A=goliu