[router] fixed memory leak issue
Recently, we found memory leak issue during load test in venice-8 of
prod-ltx1:
https://ingraphs.prod.linkedin.com/snapshot/JVM%20memory%20leak%20of%20venice-8%20in%20prod-ltx1_20180820_205921/
The issue was caused by HttpAsyncClient.
Internally, HttpAsyncClient is maintaining a LinkedList to store all the pending
requests before the downstreaming logic handles it. When the Router
is very busy, the downstreaming processing could be very slow, then
the request accumulated in the `LinkedList` could grow excessively,
which would trigger this memory leak behavior. This will form a infinite loop
since the downstreaming processing logic will be even slower because
of the increased memory usage (long GC pause).
To solve this issue, this change adds a threshold of the pending request
for each HttpAsyncClient, and Router will return 503 if the pending
request exceeds the configured threshold.
This change also adds the following metrics to track this feature:
1. `unavailable_request` per store
We should setup alert against this metric, and when it happens,
we could either bump up the configured threshold if we think there
is still capacity left or add more routers if run out of the capacity;
2. `pending_request_count`
This metric will indicate the pending request count when there is a new
request to storage node, and it will help decide whether we should
increase the configured threshold.

We verified this change in venice-8 during load test, and memory leak
behavior didn't happen. So the long-latency was caused by the router
capacity. We will tune the http client in Router to improve the
performance in the future.

RB=1402763
BUG=VOLDENG-4215
G=voldemort-reviewers
R=mwise,sdwu,yayan
A=fvillega,mnhuang