[admin tool] Restore --flat-json and --filter-json functionality to admin tool
This functionality was lost during a refactor.  Now we can use --flat-json or -fj to print
response objects as single-line json objects instead of pretty-printed with newlines.  We
can also use --filter-json or -f with a comma delimeted list of fields to filter for shorter
output.

RB=1403172
G=voldemort-reviewers
A=yayan