Add ZK status metrics to SN, router and controller
Since we heavily rely on ZK, it is very important to quickly know if
there are any issues related to ZK.

This commit adds ZK status metrics to Venice backend so that we could
set alert if ZK clients lose connections.

Unit Test:
Add tests for ZkClientStatusStats to see if it can update / emit ZK
client status properly.

RB=1404527
R=mwise,yayan,chgao,goliu
A=mwise,goliu