[common] scala cross-building
To comply with the scala cross building HI ticket (HI-66753), we need to add a scala version suffix to
the venice-common and venice-test-common modules which use scala. The scala cross-building migration
guide doesn't work for us because, in short, venice is an open-source project and livenice MP is simply
a wrapper.

RB=1406761
G=voldemort-reviewers
R=yayan
A=yayan