Fix the conflict issue between legacy topics clean-up and incremental push
Previously, parent controller would pre-check and clean up legacy topics
before creating new version. This is conflict with incremental push
where we deliberately leave the topics there and reuse them. The logic
lives in VeniceParentHelixAdmin#getTopicForCurrentPushJob.

In this RB, we let controller not delete the topic during the pre-check
if it is an ongoing incremental push. The topic would be cleaned-up when
next regluar push happens instead.

A small issue for this is that if partners trigger regular pushes but
they fail. They would not be able to do another incremental push (until
the next following regular push succeed) because the old topics have
already been deleted.

Test:
1.Add a test in HelixParentAdmin to check if we can truncate multiple
leaking topics.
2.TestMultiDataCenterPush#testMultiDataCenterIncrementalPush is back to
work again

JIRA: VOLDENG-4218

RB=1417971
R=fvillega,yayan,chgao,goliu
A=goliu