[server] fix the issue that bdb space metrics always show negative number
The negative number -16 is the NULL_BDB_ENVIRONMENT.code, which means the BDB environment is not set correctly
in the AggVersionedBdbStorageEngineStats; this is because of a bug in the `addLocalStorageEngine` method in
StoreRepository where the StorageEngine is always null when updating the stats in it.

This RB also fix a test case called TestTopicRequestOnHybridDelete.serverRestartOnHybridStoreKeepsVersionOnline
where the D2 client in SystemProducer is closed more than one time.

RB=1423265
G=voldemort-reviewers
R=sdwu,fvillega,chgao,goliu
A=fvillega