[controller] correct the sensor prefix name for AdminConsumptionStats
Context: the controller metrics called "failed_admin_messages" and "admin_message_div_error_report_count" only exist
for venice-5 cluster in prod. The root cause is that we used dot(.) in the sensor prefix name for AdminConsumptionStats:
clusterName.admin_consumption_task

In the `addMetric()` method of InGraphReporter in venice-backend, we used dot(.) as the separator for PackageName, SensorName
and AttribueName; so the SensorName is "admin_consumption_task--failed_admin_messages" and "admin_consumption_task--admin_message_div_error_report_count";
when we register a sensor for cluster venice-2, it will unregister the sensor for venice-1.

RB=1425749
G=voldemort-reviewers
R=sdwu,fvillega,chgao,goliu
A=fvillega