[server][router] Router Heartbeat health check to server actually hits SSD
This RB would make the health check request from router hit server disk, instead of responding
immediately when the router receives the check.

Another change is using the same VeniceHostHealth instance in VeniceDispatcher and the VeniceHostFinder;
previously we were using two different instance. With this change, the unhealthy host marked in the
VeniceDispatcher will not be selected in the VeniceHostFinder.

RB=1447951
G=voldemort-reviewers
R=sdwu,fvillega,xyin,chgao,goliu
A=goliu