[thin-client] log the Avro version at the beginning
There was a case where the customer build the schema with avro 1.4 and the avro dependency was specified to 1.4;
however, some other dependencies in the customer's MP brought in avro 1.7 but the customer was not aware of it.
Adding this logging will help both customer identify the actual avro version in runtime, so the customers can
build schema with the actual avro version in runtime.

RB=1455305
G=voldemort-reviewers
R=sdwu,fvillega,xyin,chgao,goliu
A=fvillega