[VOLDENG-4337][common][server] GC tuning of ingestion throttling
We moved the throttling after putting all the records returned by 'poll'
function to the buffer, and with this way, the objects in buffer of
StoreBufferService will be persisted to SSD as fast as possible. This
will reduce the 'object copy' time during GC.

Tested in venice-p, and here is the comparison:
https://ingraphs.prod.linkedin.com/snapshot/GC_metrics_after_throttling_optimization_20181031_205355/

lor1-app1442.prod.linkedin.com is with the optmization.
Based on the experiement, GC count is similar, but GC duration is reduced
by more than 70%.

Added one new config:
/**
 * This config is used to control the maximum records returned by
 * every poll request.
 * So far, Store Ingestion is throttling per poll, so if the
 * configured value is too big,
 * the throttling could be inaccurate and it may impact GC as well.
 *
 * We should try to avoid too many long-lasting objects in JVM to
 * minimize GC overhead.
 */
server.kafka.max.poll.records: default 100

Also reduced the default value for config:
store.writer.buffer.memory.capacity from 125MB to 25MB to reduce the
possiblity of keeping too many records in the buffer.

RB=1468801
BUG=VOLDENG-4337
G=voldemort-reviewers
R=sdwu,fvillega,mnhuang,chgao
A=fvillega