[VOLDENG-4350][router]Bumped up dds router dependency
When releasing 0.2.30 to prod, SRE noticed some minor latency increase
for single-get request, and batch-get latency was bad during load test.
After bumping up dds router to 5.1.10, the latency of single-get is
improved when testing in venice-p, and batch-get latency is similar.
So we decided to move forward with the latest release: 5.1.10, and if it
doesn't work in prod, we will consider to rollback to the previous
version: 4.0.144.

RB=1475484
BUG=VOLDENG-4350
G=voldemort-reviewers
R=sdwu,fvillega,mnhuang,chgao
A=fvillega