Fix Per-Store Token Bucket Metrics
The gauge-type metric used for monitoring token buckets specifies how to query the metric but not
which stores to query.  This change initializes the metric for each store and registers the metric
for store change events in order to guaranee that we emit per store token bucket metrics.

RB=1480610
G=voldemort-reviewers
R=sdwu
A=sdwu,fvillega