[test] fix a flaky test case TestRestartServerDuringIngestion
At the end of the test case, it will query all the keys from
streaming ingestion; before this check, it decides that it
could start querying if the last key from streaming ingestion
can be queried; however, this check only assures that the
partition which contains the last key is online and ready to
serve, but other partitions are not necessary online already.
The fix is to wait until all partitions are online, and then
start querying all keys from streaming ingestion.

RB=1481232
G=voldemort-reviewers
R=sdwu,fvillega,xyin,chgao,goliu
A=fvillega,xyin,goliu