[router] Router implementation for read compute
This RB includes all the changes for read compute in router.
A new VenicePath called VeniceComputePath is added; it would deserialize
the compute request and also a list of keys from client, but it would
reuse the serialized bytes for compute request in the Http request content
so that it doesn't need to serialize it again.

Most duplicate codes for both VeniceComputePath and VeniceMultiGetPath are
abstracted out to a template class called VeniceMultiKeyPath.

A new set of metrics with prefix "compute_" is added for read compute.

RB=1485168
G=voldemort-reviewers
R=sdwu,fvillega,xyin,chgao,goliu
A=goliu