[client] read computation client support
This change adds 'compute' support on client side.
Please refer to the design doc understand the high-level idea:
https://microsoft-my.sharepoint.com/:w:/g/personal/goliu_linkedin_biz/ETZiMH2BwzBCogvGan6Y-YABCPQlWboCIAs6vaqRVop2oA?e=6KALx1

So far, the generated result schema passed to backend is not optimized,
and here are the potential apporoach we could explore in the future:
 1. Remove the unnecessary fields, such as 'doc';
 2. Compress the generated result schema;
 3. The ultimate optimization: dynamic result schema registry;

We may need to add more metrics regarding to the latency of building up
the compute request later.

RB=1487490
BUG=VOLDENG-4359
G=voldemort-reviewers
R=sdwu,fvillega,mnhuang,chgao
A=fvillega,mnhuang