[thin-client] Added a new BlackHoleDeserializer, for performance testing purposes.
This BatchGetDeserializer does nothing, and always returns an empty list.

The intent is to use this serializer in order to stress the Venice backend services
with more requests, while doing as little work as possible in the client. This can
also be used to test the transport-only portion of the client, without any
deserialization costs.

RB=1497344
BUG=VOLDENG-4388
G=voldemort-reviewers
R=mnhuang
A=mnhuang