[server] Remove noisy VeniceNoHelixResourceException logging from StorageQuotaEnforcementHandler
The quota enforcement handler listens to both the external view change and the store metadata
change; both callbacks of these two changes are asynchronous so the order cannot be guaranteed.

During a new push, a new version is added to the version list of the Store metadata before the
push actually starts, so function (StorageQuotaEnforcementHandler#handleStoreChanged()) is
invoked before the new resource assignment shows up in the external view, so calling
routingRepository.getPartitionAssignments() for a the future version will fail in most cases,
because the new topic is not in the external view at all.

RB=1519172
BUG=VOLDENG-4414
G=voldemort-reviewers
R=sdwu,fvillega,chgao,goliu
A=fvillega