[client] Support cosine similarity and allow list as input in compute API
This RB adds support for cosine similarity on the client side. Besides,
Collection type is supported in projection API and List type is supported
in DotProduct API.

Fix a bug in AvroComputeRequestBuilder; the error map field can not be
static because the position metadata of the field would be updated after
setting the error map field in a schema.

RB=1523825
BUG=VOLDENG-4430
G=voldemort-reviewers
R=sdwu,fvillega,xyin,chgao,goliu
A=fvillega,goliu