[server] Introduced a new ComputablePrimitiveFloatList, used by read computation.
This is a mash up of Avro's GenericData.Array class, as well as Quasar's cacheable computation.

It is used by the Venice server to provide efficient deserialization and handling of float lists in
read computation use cases.

Compared to the Avro implementation, it offers the following GC-related optimizations:

- It does not, by default, box primitive floats into Object Floats, though it will still do so if the
  regular functions are called (e.g.: #get()), for compatibility purposes. In order to avoid boxing,
  the #getPrimitive() function can be used instead.
- It does not maintain a reference to a Schema instance, since that schema would always be the same.
  Instead, it defines a static #SCHEMA which is used by all instances.
- It re-implements #compareTo(), #equals() and #hashCode() in order to leverage the primitive types,
  rather than causing unintended boxing.

It also offers a #squaredL2Norm() function, which internally caches its results so that it avoids
recomputing the same thing repeatedly. This functionality comes at the expense of an extra float and
boolean properties, which adds up to less than the overhead saved by the object minimization described
above.

RB=1535288
BUG=VOLDENG-4469
G=voldemort-reviewers
R=sdwu,mnhuang,xyin,chgao,goliu
A=sdwu,mnhuang